Identifier renaming cleanup.

+++ lzw_decompress.h
@@ -6,8 +6,8 @@
 struct lzwd_state {
 	uint32_t code_width;
 	uint32_t code_mask;
-	uint32_t next_code;
-	uint32_t old_code;
+	uint16_t next_code;
+	uint16_t old_code;
 	uint32_t tree[LZW_MAX_CODE]; // 16K
 
 	bool	 must_reset;
